Item型の修正
