+++ src/logics/circles/index.ts
@@ -16,6 +16,9 @@ export const useGetCircleById = (id: number) => {
               eventName: "技術書典6",
               publishedAt: new Date("2018/04/24"),
               path: "path",
+              numberOfPages: 122,
+              price: 1_000,
+              tags: ["勉強会", "成長", "勉強会運営", "勉強会参加"],
             },
           ],
         },
