--- src/types.ts
@@ -1,8 +1,0 @@
-export type Item = {
-  category: string;
-  imagePath: string;
-  title: string;
-  eventName?: string;
-  publishedAt: Date;
-  path: string;
-};
